unit tests for visualizeStateMachineCDK (draft)
