--- src/test/cdk/explorer/commands/visualizeStateMachineCDK2.test.ts
@@ -1,489 +1,0 @@
-/*!
- * Copyright 2018-2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
- * SPDX-License-Identifier: Apache-2.0
- */
-
-import * as assert from 'assert'
-import * as sinon from 'sinon'
-import * as vscode from 'vscode'
-
-import { Disposable } from 'vscode-languageclient'
-import { AslVisualizationCDK } from '../../../../../src/cdk/commands/aslVisualizationCDK'
-import { AslVisualizationCDKManager, getCDKAppName } from '../../../../../src/cdk/commands/aslVisualizationCDKManager'
-
-import { ext } from '../../../../shared/extensionGlobals'
-import { StateMachineGraphCache } from '../../../../stepFunctions/utils'
-
-import { JSON_ASL } from '../../../../../src/stepFunctions/constants/aslFormats'
-import { ConstructTreeEntity } from '../../../../../src/cdk/explorer/tree/types'
-import { ConstructNode, isStateMachine } from '../../../../../src/cdk/explorer/nodes/constructNode'
-import { FakeParentNode } from '../constructNode.test'
-import * as treeUtils from '../../utilities/treeTestUtils'
-import { getLogger, Logger } from '../../../../shared/logger'
-
-
-
-// Top level defintions
-let mockAslVisualizationCDKManager: MockAslVisualizationCDKManager
-let sandbox: sinon.SinonSandbox
-
-const mockGlobalStorage: vscode.Memento = {
-    update: sinon.spy(),
-    get: sinon.stub().returns(undefined),
-}
-
-const mockUriOne: vscode.Uri = {
-    authority: 'amazon.com',
-    fragment: 'MockFragmentOne',
-    fsPath: 'MockFSPathOne',
-    query: 'MockQueryOne',
-    path: '/MockPathOne',
-    scheme: 'MockSchemeOne',
-    with: () => {
-        return mockUriOne
-    },
-    toJSON: sinon.spy(),
-}
-
-const mockTextDocumentOne: vscode.TextDocument = {
-    eol: 1,
-    fileName: 'MockFileNameOne',
-    isClosed: false,
-    isDirty: false,
-    isUntitled: false,
-    languageId: 'MockLanguageIdOne',
-    lineCount: 0,
-    uri: mockUriOne,
-    version: 0,
-    getText: () => {
-        return 'MockDocumentTextOne'
-    },
-    getWordRangeAtPosition: sinon.spy(),
-    lineAt: sinon.spy(),
-    offsetAt: sinon.spy(),
-    positionAt: sinon.spy(),
-    save: sinon.spy(),
-    validatePosition: sinon.spy(),
-    validateRange: sinon.spy(),
-}
-
-const mockUriTwo: vscode.Uri = {
-    authority: 'amazon.org',
-    fragment: 'MockFragmentTwo',
-    fsPath: 'MockFSPathTwo',
-    query: 'MockQueryTwo',
-    path: '/MockPathTwo',
-    scheme: 'MockSchemeTwo',
-    with: sinon.spy(),
-    toJSON: sinon.spy(),
-}
-
-const mockTextDocumentTwo: vscode.TextDocument = {
-    eol: 1,
-    fileName: 'MockFileNameTwo',
-    isClosed: false,
-    isDirty: false,
-    isUntitled: false,
-    languageId: 'MockLanguageIdTwo',
-    lineCount: 0,
-    uri: mockUriTwo,
-    version: 0,
-    getText: () => {
-        return 'MockDocumentTextTwo'
-    },
-    getWordRangeAtPosition: sinon.spy(),
-    lineAt: sinon.spy(),
-    offsetAt: sinon.spy(),
-    positionAt: sinon.spy(),
-    save: sinon.spy(),
-    validatePosition: sinon.spy(),
-    validateRange: sinon.spy(),
-}
-
-const mockUriThree: vscode.Uri = {
-    authority: 'amazon.de',
-    fragment: 'MockFragmentYaml',
-    fsPath: 'MockFSPathYaml',
-    query: 'MockQueryYaml',
-    path: '/MockPathYaml',
-    scheme: 'MockSchemeYaml',
-    with: sinon.spy(),
-    toJSON: sinon.spy(),
-}
-
-const mockDataJson =
-    '{"Comment":"A Hello World example of the Amazon States Language using Pass states","StartAt":"Hello","States":{"Hello":{"Type":"Pass","Result":"Hello","Next":"World"},"World":{"Type":"Pass","Result":"${Text}","End":true}}}'
-
-
-const mockTextDocumentJson: vscode.TextDocument = {
-    eol: 1,
-    fileName: 'MockFileNameJson',
-    isClosed: false,
-    isDirty: false,
-    isUntitled: false,
-    languageId: JSON_ASL,
-    lineCount: 0,
-    uri: mockUriThree,
-    version: 0,
-    getText: () => {
-        return mockDataJson
-    },
-    getWordRangeAtPosition: sinon.spy(),
-    lineAt: sinon.spy(),
-    offsetAt: sinon.spy(),
-    positionAt: sinon.spy(),
-    save: sinon.spy(),
-    validatePosition: sinon.spy(),
-    validateRange: sinon.spy(),
-}
-
-const mockExtensionContext: vscode.ExtensionContext = {
-    extensionPath: '',
-    globalState: mockGlobalStorage,
-    globalStoragePath: '',
-    logPath: '',
-    storagePath: '',
-    subscriptions: [],
-    workspaceState: mockGlobalStorage,
-    asAbsolutePath: sinon.spy(),
-}
-
-const mockNonSMConstructTreeEntity: ConstructTreeEntity = {
-    id: 'MyLambdaFunction',
-    path: 'aws-tester/MyLambdaFunction',
-    children: {
-        'Resource': {
-            id: 'Resource',
-            path: 'aws-tester/MyLambdaFunction/Resource',
-            attributes: {
-                "aws:cdk:cloudformation:type": 'AWS::StepFunctions::LambdaFunction'
-            }
-        }
-    }
-}
-
-const mockNonConstructNode = new ConstructNode(
-    new FakeParentNode('cdkJsonPath'),
-    'MyLambdaFunction',
-    vscode.TreeItemCollapsibleState.Collapsed,
-    mockNonSMConstructTreeEntity
-)
-
-const mockSMConstructTreeEntity: ConstructTreeEntity = {
-    id: 'MyStateMachine',
-    path: 'aws-tester/MyStateMachine',
-    children: {
-        'Resource': {
-            id: 'Resource',
-            path: 'aws-tester/MyStateMachine/Resource',
-            attributes: {
-                "aws:cdk:cloudformation:type": 'AWS::StepFunctions::StateMachine'
-            }
-        }
-    }
-}
-
-const mockStateMachineNode = new ConstructNode(
-    new FakeParentNode('cdkJsonPath'),
-    'MyStateMachine',
-    vscode.TreeItemCollapsibleState.Collapsed,
-    mockSMConstructTreeEntity
-)
-
-const mockStateMachineNode2 = new ConstructNode(
-    new FakeParentNode('cdkJsonPath'),
-    'MyStateMachine2',
-    vscode.TreeItemCollapsibleState.Collapsed,
-    mockSMConstructTreeEntity
-)
-
-// const mockNonConstructNode2: ConstructNode= {
-//     type: 'ConstructNode',
-//     properties: undefined,
-//     parent: new FakeParentNode('cdkJsonPath'),
-//     label: 'MyLambdaFunction',
-//     collapsibleState: vscode.TreeItemCollapsibleState.Collapsed,
-//     construct: treeUtils.generateConstructTreeEntity('MyLambdaFunction', 'constructTreePath', true),
-//     contextValue: 'awsCdkConstructNode',
-//     getChildren: sinon.spy()
-// }
-
-describe('StepFunctions VisualizeStateMachine', async function () {
-    //let mockVsCode: MockVSCode
-
-    const oldWebviewScriptsPath = ext.visualizationResourcePaths.localWebviewScriptsPath
-    const oldWebviewBodyPath = ext.visualizationResourcePaths.webviewBodyScript
-    const oldCachePath = ext.visualizationResourcePaths.visualizationLibraryCachePath
-    const oldScriptPath = ext.visualizationResourcePaths.visualizationLibraryScript
-    const oldCssPath = ext.visualizationResourcePaths.visualizationLibraryCSS
-    const oldThemePath = ext.visualizationResourcePaths.stateMachineCustomThemePath
-    const oldThemeCssPath = ext.visualizationResourcePaths.stateMachineCustomThemeCSS
-
-    // Before all
-    before(function () {
-        //mockVsCode = new MockVSCode()
-
-        ext.visualizationResourcePaths.localWebviewScriptsPath = mockUriOne
-        ext.visualizationResourcePaths.visualizationLibraryCachePath = mockUriOne
-        ext.visualizationResourcePaths.stateMachineCustomThemePath = mockUriOne
-        ext.visualizationResourcePaths.webviewBodyScript = mockUriOne
-        ext.visualizationResourcePaths.visualizationLibraryScript = mockUriOne
-        ext.visualizationResourcePaths.visualizationLibraryCSS = mockUriOne
-        ext.visualizationResourcePaths.stateMachineCustomThemeCSS = mockUriOne
-
-        sandbox = sinon.createSandbox()
-        sandbox.stub(StateMachineGraphCache.prototype, 'updateCachedFile').callsFake(async options => {
-            return
-        })
-    })
-
-    // Before each
-    beforeEach(function () {
-        mockAslVisualizationCDKManager = new MockAslVisualizationCDKManager(mockExtensionContext)
-    })
-
-    // After each
-    afterEach(function () {
-        //mockVsCode.closeAll()
-    })
-
-    // After all
-    after(function () {
-        sandbox.restore()
-        ext.visualizationResourcePaths.localWebviewScriptsPath = oldWebviewScriptsPath
-        ext.visualizationResourcePaths.webviewBodyScript = oldWebviewBodyPath
-        ext.visualizationResourcePaths.visualizationLibraryCachePath = oldCachePath
-        ext.visualizationResourcePaths.visualizationLibraryScript = oldScriptPath
-        ext.visualizationResourcePaths.visualizationLibraryCSS = oldCssPath
-        ext.visualizationResourcePaths.stateMachineCustomThemePath = oldThemePath
-        ext.visualizationResourcePaths.stateMachineCustomThemeCSS = oldThemeCssPath
-    })
-
-    // Tests
-    it('Test AslVisualizationCDK on setup all properties are correct', function () {
-        const vis = new MockAslVisualizationCDK(mockTextDocumentOne, '', '')
-
-        assert.deepStrictEqual(vis.documentUri, mockTextDocumentOne.uri)
-        assert.strictEqual(vis.getIsPanelDisposed(), false)
-        assert.strictEqual(vis.getDisposables().length, 5)
-
-        let panel = vis.getPanel()
-        assert.ok(panel)
-        panel = panel as vscode.WebviewPanel
-        assert.ok(panel.title.length > 0)
-        assert.strictEqual(panel.viewType, 'stateMachineVisualization')
-
-        let webview = vis.getWebview()
-        assert.ok(webview)
-        webview = webview as vscode.Webview
-        assert.ok(webview.html)
-    })
-
-    it('Test AslVisualizationCDKManager on setup managedVisualizationsCDK set is empty', function () {
-        assert.strictEqual(mockAslVisualizationCDKManager.getManagedVisualizations().size, 0)
-    })
-
-    it('Test AslVisualizationCDKManager managedVisualizationsCDK set still empty if node is not of type state machine', async function () {
-        assert.strictEqual(mockAslVisualizationCDKManager.getManagedVisualizations().size, 0)
-
-        // Preview with non state machine node
-        assert.strictEqual(await mockAslVisualizationCDKManager.visualizeStateMachine(mockGlobalStorage, mockNonConstructNode), undefined)
-        assert.strictEqual(mockAslVisualizationCDKManager.getManagedVisualizations().size, 0)
-    })
-
-    it('Test AslVisualizationCDKManager managedVisualizationsCDK set has one AslVisCDK on first visualization', async function () {
-        assert.strictEqual(mockAslVisualizationCDKManager.getManagedVisualizations().size, 0)
-
-        //assert.ok(mockStateMachineNode.contextValue === 'awsCdkStateMachineNode')
-        assert.ok(await mockAslVisualizationCDKManager.visualizeStateMachine(mockGlobalStorage, mockStateMachineNode))
-        assert.strictEqual(mockAslVisualizationCDKManager.getManagedVisualizations().size, 1)
-
-        //const managedVisualizationsCDK = mockAslVisualizationCDKManager.getManagedVisualizations()
-        //assert.ok(managedVisualizationsCDK.get(mockTextDocumentOne.uri.path))
-    })
-
-    it('Test AslVisualizationCDKManager managedVisualizationsCDK set does not add second VisCDK on duplicate state machine node', async function () {
-        assert.strictEqual(mockAslVisualizationCDKManager.getManagedVisualizations().size, 0)
-
-        // visualization for mockStateMachineNode
-        await mockAslVisualizationCDKManager.visualizeStateMachine(mockGlobalStorage, mockStateMachineNode)
-        assert.strictEqual(mockAslVisualizationCDKManager.getManagedVisualizations().size, 1)
-
-        // visualization for mockStateMachineNode again
-        await mockAslVisualizationCDKManager.visualizeStateMachine(mockGlobalStorage, mockStateMachineNode)
-        assert.strictEqual(mockAslVisualizationCDKManager.getManagedVisualizations().size, 1)
-
-        //const managedVisualizations = aslVisualizationCDKManager.getManagedVisualizations()
-        //assert.ok(managedVisualizations.get(mockTextDocumentOne.uri.path))
-    })
-
-    it('Test AslVisualizationCDKManager managedVisualizationsCDK set adds second VisCDK on different state machine nodes', async function () {
-        assert.strictEqual(mockAslVisualizationCDKManager.getManagedVisualizations().size, 0)
-
-        // visualization for mockStateMachineNode
-        await mockAslVisualizationCDKManager.visualizeStateMachine(mockGlobalStorage, mockStateMachineNode)
-        assert.strictEqual(mockAslVisualizationCDKManager.getManagedVisualizations().size, 1)
-
-        // visualization for mockStateMachineNode2
-        await mockAslVisualizationCDKManager.visualizeStateMachine(mockGlobalStorage, mockStateMachineNode2)
-        assert.strictEqual(mockAslVisualizationCDKManager.getManagedVisualizations().size, 2)
-
-        // const managedVisualizations = aslVisualizationCDKManager.getManagedVisualizations()
-        // assert.ok(managedVisualizations.get(mockTextDocumentOne.uri.path))
-        // assert.ok(managedVisualizations.get(mockTextDocumentTwo.uri.path))
-    })
-
-    it('Test AslVisualizationCDKManager managedVisualizationsCDK set does not add duplicate renders when multiple VisCDK active', async function () {
-        assert.strictEqual(mockAslVisualizationCDKManager.getManagedVisualizations().size, 0)
-
-        // visualization for mockStateMachineNode
-        await mockAslVisualizationCDKManager.visualizeStateMachine(mockGlobalStorage, mockStateMachineNode)
-        assert.strictEqual(mockAslVisualizationCDKManager.getManagedVisualizations().size, 1)
-
-        // visualization for mockStateMachineNode2
-        await mockAslVisualizationCDKManager.visualizeStateMachine(mockGlobalStorage, mockStateMachineNode2)
-        assert.strictEqual(mockAslVisualizationCDKManager.getManagedVisualizations().size, 2)
-
-        // visualization for mockStateMachineNode again
-        await mockAslVisualizationCDKManager.visualizeStateMachine(mockGlobalStorage, mockStateMachineNode)
-        assert.strictEqual(mockAslVisualizationCDKManager.getManagedVisualizations().size, 2)
-
-        // visualization for mockStateMachineNode2 again
-        await mockAslVisualizationCDKManager.visualizeStateMachine(mockGlobalStorage, mockStateMachineNode2)
-        assert.strictEqual(mockAslVisualizationCDKManager.getManagedVisualizations().size, 2)
-
-        // const managedVisualizations = aslVisualizationCDKManager.getManagedVisualizations()
-        // assert.ok(managedVisualizations.get(mockTextDocumentOne.uri.path))
-        // assert.ok(managedVisualizations.get(mockTextDocumentTwo.uri.path))
-    })
-
-    // it('Test AslVisualizationCDKManager managedVisualizationsCDK set removes visualization on removal of template.json file, single visCDK', async function () {
-    //     assert.strictEqual(mockAslVisualizationCDKManager.getManagedVisualizations().size, 0)
-
-    //     // Preview Doc1
-    //     mockVsCode.showTextDocument(mockTextDocumentOne)
-    //     let panel = await aslVisualizationCDKManager.visualizeStateMachine(
-    //         mockGlobalStorage,
-    //         vscode.window.activeTextEditor
-    //     )
-    //     assert.strictEqual(aslVisualizationCDKManager.getManagedVisualizations().size, 1)
-
-    //     // Dispose of visualization panel
-    //     assert.ok(panel, 'Panel was not successfully generated')
-    //     panel = panel as vscode.WebviewPanel
-    //     panel.dispose()
-
-    //     assert.strictEqual(aslVisualizationCDKManager.getManagedVisualizations().size, 0)
-    // })
-
-    // it('Test AslVisualizationCDKManager managedVisualizationsCDK set removes correct visualization on removal of template.json file, multiple vis', async function () {
-    //     assert.strictEqual(aslVisualizationCDKManager.getManagedVisualizations().size, 0)
-
-    //     // Preview Doc1
-    //     mockVsCode.showTextDocument(mockTextDocumentOne)
-    //     let panel = await aslVisualizationCDKManager.visualizeStateMachine(
-    //         mockGlobalStorage,
-    //         vscode.window.activeTextEditor
-    //     )
-    //     assert.strictEqual(aslVisualizationCDKManager.getManagedVisualizations().size, 1)
-
-    //     // Preview Doc2
-    //     mockVsCode.showTextDocument(mockTextDocumentTwo)
-    //     await aslVisualizationCDKManager.visualizeStateMachine(mockGlobalStorage, vscode.window.activeTextEditor)
-    //     assert.strictEqual(aslVisualizationCDKManager.getManagedVisualizations().size, 2)
-
-    //     // Dispose of first visualization panel
-    //     assert.ok(panel, 'Panel was not successfully generated')
-    //     panel = panel as vscode.WebviewPanel
-    //     panel.dispose()
-
-    //     assert.strictEqual(aslVisualizationCDKManager.getManagedVisualizations().size, 1)
-    // })
-
-    // it('Test AslVisualisationCDK sendUpdateMessage posts a correct update message for ASL files', async function () {
-    //     const postMessage = sinon.spy()
-    //     class MockAslVisualizationJsonCDK extends AslVisualizationCDK {
-    //         public getWebview(): vscode.Webview | undefined {
-    //             return ({ postMessage } as unknown) as vscode.Webview
-    //         }
-    //     }
-
-    //     const visualisation = new MockAslVisualizationJsonCDK(mockTextDocumentJson,'','')
-
-    //     await visualisation.sendUpdateMessage(mockTextDocumentJson)
-
-    //     const message = {
-    //         command: 'update',
-    //         stateMachineData: mockDataJson,
-    //         isValid: true,
-    //         errors: [],
-    //     }
-
-    //     assert.ok(postMessage.calledOnce)
-    //     assert.deepEqual(postMessage.firstCall.args, [message])
-    // })
-})
-
-class MockAslVisualizationCDK extends AslVisualizationCDK {
-    protected getText(textDocument: vscode.TextDocument): string {
-        return mockDataJson
-    }
-
-    protected getTemplateJsonDocument(templatePath: string) {
-        return mockUriOne
-    }
-
-    public getIsPanelDisposed(): boolean {
-        return this.isPanelDisposed
-    }
-
-    public getDisposables(): Disposable[] {
-        return this.disposables
-    }
-}
-
-class MockAslVisualizationCDKManager extends AslVisualizationCDKManager {
-    public async visualizeStateMachine(
-        globalStorage: vscode.Memento,
-        node: ConstructNode
-    ): Promise<vscode.WebviewPanel | undefined> {
-        if (!isStateMachine(node.construct)) {
-            console.log('is not a state machine')
-            return
-        }
-
-        const logger: Logger = getLogger()
-        const cache = new StateMachineGraphCache()
-        const uniqueIdentifier = node.label
-        const cdkOutPath = node.id?.replace(`/tree.json/${node.tooltip}`, ``)
-        const stackName = node.tooltip?.replace(`/${uniqueIdentifier}`, ``)
-        const templatePath = String(cdkOutPath) + `/${stackName}.template.json`
-        const appName = getCDKAppName(cdkOutPath!)
-        const existingVisualization = this.getExistingVisualization(appName, uniqueIdentifier)
-        const uri = vscode.Uri.file(templatePath);
-
-        if (existingVisualization) {
-            existingVisualization.showPanel()
-
-            return existingVisualization.getPanel()
-        }
-
-        // Existing visualization does not exist, construct new visualization
-        try {
-            //await cache.updateCache(globalStorage)
-
-            //const textDocument = await vscode.workspace.openTextDocument(uri)
-            const newVisualization = new MockAslVisualizationCDK(mockTextDocumentOne, templatePath, uniqueIdentifier)
-            if (newVisualization) {
-                this.handleNewVisualization(appName, newVisualization)
-                return newVisualization.getPanel()
-            }
-        } catch (err) {
-            this.handleErr(err, logger)
-        }
-
-        console.log('cannot render visualization')
-        return
-    }
-}
-
